Completely change Client struct and add command system
Sharing the state of the Client object was becoming a challenge,
using a Mutex would've resulted in far too much deadlocking and
would've been way too slow. Instead, channels were used to share the
state of the internal ClientData struct.

The owner of the "true" client is the task that processes the packets
and is also tasked with handling messages defined by ClientCommand.

Currently, the only commands are a basic Ping and a simple mockup of
how a event system would work with the OnLogin command. To send the
message from the public Client to the internal client a mpsc channel
is used and for commands that expect an output a oneshot channel is
established to return that result.

In the case of the event handling commands, such as OnLogin a watcher
channel is created to have a single producer multi consumer
architecture.